refactor(login): move `auth.js`, `passportFacebook.js`
OAuth service provider(facebook, google)가 추가됨에 따라, auth라는 폴더로 그룹화함.
또한 `passportFacebook.js`에서 `passport.js`로 수정함.
