+++ back/package.json
@@ -36,6 +36,7 @@
     "mysql2": "^1.7.0",
     "passport": "^0.4.0",
     "passport-facebook": "^3.0.0",
+    "passport-google-oauth": "^2.0.0",
     "sequelize": "^5.19.6",
     "sequelize-cli": "^5.5.1"
   },
